Move `symbol_table_parser` under `Tapioca::Static`
