+++ lib/tapioca/internal.rb
@@ -15,7 +15,7 @@ require "tapioca/commands"
 require "tapioca/cli"
 require "tapioca/gemfile"
 require "tapioca/executor"
-require "tapioca/symbol_table_parser"
+require "tapioca/static/symbol_table_parser"
 require "tapioca/static/symbol_loader"
 require "tapioca/gem/events"
 require "tapioca/gem/listeners"
